+++ base/associative.jl
@@ -317,7 +317,7 @@ end
     filter!(f, d::Associative)
 
 Update `d`, removing elements for which `f` is `false`.
-The function `f` is passed two arguments (key and value).
+The function `f` is passed `key=>value` pairs.
 
 # Example
 ```jldoctest
@@ -327,7 +327,7 @@ Dict{Int64,String} with 3 entries:
   3 => "c"
   1 => "a"
 
-julia> filter!((x,y)->isodd(x), d)
+julia> filter!(p->isodd(p.first), d)
 Dict{Int64,String} with 2 entries:
   3 => "c"
   1 => "a"
@@ -335,13 +335,35 @@ Dict{Int64,String} with 2 entries:
 """
 function filter!(f, d::Associative)
     badkeys = Vector{keytype(d)}(0)
-    for (k,v) in d
-        # don't delete!(d, k) here, since associative types
-        # may not support mutation during iteration
-        f(k,v) || push!(badkeys, k)
+    try
+        for (k,v) in d
+            # don't delete!(d, k) here, since associative types
+            # may not support mutation during iteration
+            f(k => v) || push!(badkeys, k)
+        end
+    catch e
+        return filter!_dict_deprecation(e, f, d)
     end
     for k in badkeys
         delete!(d, k)
+    end
+    return d
+end
+
+function filter!_dict_deprecation(e, f, d::Associative)
+    if isa(e, MethodError) && e.f === f
+        depwarn("In `filter!(f, dict)`, `f` is now passed a single pair instead of two arguments.", :filter!)
+        badkeys = Vector{keytype(d)}(0)
+        for (k,v) in d
+            # don't delete!(d, k) here, since associative types
+            # may not support mutation during iteration
+            f(k, v) || push!(badkeys, k)
+        end
+        for k in badkeys
+            delete!(d, k)
+        end
+    else
+        rethrow(e)
     end
     return d
 end
@@ -350,7 +372,7 @@ end
     filter(f, d::Associative)
 
 Return a copy of `d`, removing elements for which `f` is `false`.
-The function `f` is passed two arguments (key and value).
+The function `f` is passed `key=>value` pairs.
 
 # Examples
 ```jldoctest
@@ -359,7 +381,7 @@ Dict{Int64,String} with 2 entries:
   2 => "b"
   1 => "a"
 
-julia> filter((x,y)->isodd(x), d)
+julia> filter(p->isodd(p.first), d)
 Dict{Int64,String} with 1 entry:
   1 => "a"
 ```
@@ -367,9 +389,22 @@ Dict{Int64,String} with 1 entry:
 function filter(f, d::Associative)
     # don't just do filter!(f, copy(d)): avoid making a whole copy of d
     df = similar(d)
-    for (k,v) in d
-        if f(k,v)
-            df[k] = v
+    try
+        for (k, v) in d
+            if f(k => v)
+                df[k] = v
+            end
+        end
+    catch e
+        if isa(e, MethodError) && e.f === f
+            depwarn("In `filter(f, dict)`, `f` is now passed a single pair instead of two arguments.", :filter)
+            for (k, v) in d
+                if f(k, v)
+                    df[k] = v
+                end
+            end
+        else
+            rethrow(e)
         end
     end
     return df