+++ base/deprecated.jl
@@ -1687,6 +1687,9 @@ end
 # issue #5148, PR #23259
 # warning for `const` on locals should be changed to an error in julia-syntax.scm
 
+# issue #17886
+# deprecations for filter[!] with 2-arg functions are in associative.jl
+
 # END 0.7 deprecations
 
 # BEGIN 1.0 deprecations